move ampl examples into example feeds
