--- src/telliot_ampl_feeds/__init__.py
@@ -1,1 +1,0 @@
-"""AMPL/USD/VWAP plugins."""
