--- src/telliot_ampl_feeds/feeds/__init__.py
@@ -1,1 +1,0 @@
-"""Telliot AMPL feeds"""
