--- src/telliot_ampl_feeds/reporters/__init__.py
@@ -1,1 +1,0 @@
-"""Example AMPL Reporters."""
