--- src/telliot_ampl_feeds/scripts/__init__.py
@@ -1,1 +1,0 @@
-"""Scripts for reporting values using AMPL datafeeds."""
