--- src/telliot_ampl_feeds/sources/__init__.py
@@ -1,1 +1,0 @@
-"""Telliot AMPL sources"""
