+++ src/telliot_feed_examples/config/ampl.py
