+++ src/telliot_feed_examples/feeds/ampl_usd_vwap_feed.py
