+++ src/telliot_feed_examples/feeds/btc_usd_feed.py
@@ -1,7 +1,6 @@
 """Example datafeed used by BTCUSDReporter."""
 from telliot_core.datafeed import DataFeed
 from telliot_core.queries.coin_price import CoinPrice
-
 from telliot_feed_examples.sources.bittrex import BittrexPriceSource
 from telliot_feed_examples.sources.coinbase import CoinbasePriceSource
 from telliot_feed_examples.sources.coingecko import CoinGeckoPriceSource
