+++ src/telliot_feed_examples/feeds/uspce_feed.py
