+++ src/telliot_feed_examples/sources/ampl_usd_vwap.py
