+++ src/telliot_feed_examples/sources/uspce.py
