+++ src/telliot_feed_examples/utils/profitcalc.py
