+++ tests/feeds/test_ampl_usd_vwap_feed.py
