+++ tests/feeds/test_btc_usd_feed.py
