+++ tests/reporters/test_ampl_usd_vwap_reporter.py
