+++ tests/reporters/test_interval_reporter.py
