+++ tests/sources/test_ampl_usd_vwap_source.py
